:white_check_mark: test: 실패하는 테스트 수정
