+++ src/test/java/tipitapi/drawmytoday/domain/diary/controller/ImageControllerTest.java
@@ -2,15 +2,21 @@ package tipitapi.drawmytoday.domain.diary.controller;
 
 import static org.mockito.Mockito.verify;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import java.util.HashMap;
+import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
 import org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors;
 import org.springframework.test.web.servlet.ResultActions;
 import tipitapi.drawmytoday.common.controller.ControllerTestSetup;
@@ -45,6 +51,52 @@ class ImageControllerTest extends ControllerTestSetup {
             // then
             result.andExpect(status().isNoContent());
             verify(imageService).deleteImage(imageId, REQUEST_USER_ID);
+        }
+    }
+
+    @Nested
+    @DisplayName("reviewImage 메서드는")
+    class ReviewImageTest {
+
+        @ParameterizedTest
+        @ValueSource(strings = {"", "0", "6", "a"})
+        @DisplayName("review 값이 없거나 1~5 사이의 숫자가 아니면 BAD_REQUEST 상태코드를 응답한다.")
+        void invalid_request_body_then_return_400(String review) throws Exception {
+            // given
+            Long imageId = 1L;
+
+            // when
+            Map<String, Object> requestMap = new HashMap<>();
+            requestMap.put("review", review);
+            String requestBody = objectMapper.writeValueAsString(requestMap);
+            ResultActions result = mockMvc.perform(post(BASIC_URL + "/" + imageId + "/review")
+                .with(SecurityMockMvcRequestPostProcessors.csrf())
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(requestBody));
+
+            // then
+            result.andExpect(status().isBadRequest());
+        }
+
+        @ParameterizedTest
+        @ValueSource(strings = {"1", "2", "3", "4", "5"})
+        @DisplayName("review에 1~5 사이의 값이 들어오면 diaryId에 해당하는 일기를 리뷰한다.")
+        void review_diary(String review) throws Exception {
+            // given
+            Long imageId = 1L;
+
+            // when
+            Map<String, Object> requestMap = new HashMap<>();
+            requestMap.put("review", review);
+            String requestBody = objectMapper.writeValueAsString(requestMap);
+            ResultActions result = mockMvc.perform(post(BASIC_URL + "/" + imageId + "/review")
+                .with(SecurityMockMvcRequestPostProcessors.csrf())
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(requestBody));
+
+            // then
+            result.andExpect(status().isNoContent());
+            verify(imageService).reviewDiary(imageId, REQUEST_USER_ID, review);
         }
     }
 }
